Name refactor. Decorator refactor
Now the decorator has been divided into two files, the decorator and the post processor
