+++ framework/wazuh/rbac/pre_processor.py
